+++ ui/src/app.ts
@@ -1,10 +1,9 @@
 import { platforma } from "@platforma-open/milaboratories.mixcr-amplicon-alignment.model";
-import { defineApp } from "@platforma-sdk/ui-vue";
+import { defineAppV3 } from "@platforma-sdk/ui-vue";
 import MainPage from "./pages/MainPage.vue";
 import QcReportTablePage from "./pages/QcReportTablePage.vue";
-import { watch } from "vue";
 
-export const sdkPlugin = defineApp(platforma, () => {
+export const sdkPlugin = defineAppV3(platforma, () => {
   return {
     routes: {
       "/": () => MainPage,
@@ -14,12 +13,3 @@ export const sdkPlugin = defineApp(platforma, () => {
 });
 
 export const useApp = sdkPlugin.useApp;
-
-// Make sure labels are initialized
-const unwatch = watch(sdkPlugin, ({ loaded }) => {
-  if (!loaded) return;
-  const app = useApp();
-  app.model.args.customBlockLabel ??= "";
-  app.model.args.defaultBlockLabel ??= "Select Clonotype Definition";
-  unwatch();
-});