+++ ui/src/pages/QcReportTablePage.vue
@@ -13,7 +13,7 @@ const tableSettings = usePlDataTableSettingsV2({
   <PlBlockPage>
     <template #title> QC Report Table </template>
     <PlAgDataTableV2
-      v-model="app.model.ui.tableState"
+      v-model="app.model.data.tableState"
       :settings="tableSettings"
       show-columns-panel
       show-export-button